reactive_tui::events::actions

Struct ActionDispatcher

pub struct ActionDispatcher { /* private fields */ }

Action dispatcher that manages and routes actions

Implementations

impl ActionDispatcher

pub fn new() -> Self

Create a new action dispatcher

pub fn register<F>(&mut self, action_name: &str, handler: F)
where F: Fn(&mut Action) -> ActionResult + Send + Sync + 'static,

Register an action handler

pub fn register_multiple<F>(&mut self, action_names: &[&str], handler: F)
where F: Fn(&mut Action) -> ActionResult + Send + Sync + 'static + Clone,

Register multiple actions with the same handler

pub fn unregister(&mut self, action_name: &str)

Unregister an action handler

pub fn is_registered(&self, action_name: &str) -> bool

Check if an action is registered

pub fn dispatch(&self, action: Action) -> ActionResult

Dispatch an action immediately

pub fn send(&self, action: Action) -> Result<()>

Send an action to be processed asynchronously

pub fn take_receiver(&mut self) -> Option<UnboundedReceiver<Action>>

Get the action receiver for processing actions in an event loop

pub fn set_hierarchy(&self, hierarchy: Vec<String>) -> Result<()>

Set up action hierarchy for bubbling actions up the component tree

pub fn add_to_hierarchy(&self, element_id: String) -> Result<()>

Add element to action hierarchy

pub fn remove_from_hierarchy(&self, element_id: &str) -> Result<bool>

Remove element from action hierarchy

pub fn get_hierarchy(&self) -> Result<Vec<String>>

Get current action hierarchy

pub fn get_registered_actions(&self) -> Vec<String>

Get a list of all registered action names

pub fn action<S: Into<String>>(&self, name: S) -> ActionBuilder

Create an action builder for fluent API