reactive_tui::events::messages

Struct MessageManager

pub struct MessageManager { /* private fields */ }

Manager for the message-based event system

Implementations

impl MessageManager

pub fn new() -> Self

Create a new message manager

pub fn on<T, F>(&self, handler: F) -> Result<()>
where T: Message + 'static, F: Fn(&mut MessageEvent) -> Result<()> + Send + Sync + 'static,

Register a global message handler for a specific message type

pub fn on_element<T, F>(&self, element_id: &str, handler: F) -> Result<()>
where T: Message + 'static, F: Fn(&mut MessageEvent) -> Result<()> + Send + Sync + 'static,

Register a message handler for a specific element

pub fn send_from( &self, element_id: Option<String>, message: impl Message, ) -> Result<()>

Send a message from a specific element

pub fn send(&self, message: impl Message) -> Result<()>

Send a message (convenience method)

pub fn process_message(&self, event: MessageEvent) -> Result<()>

Process a message event, including bubbling

pub fn update_hierarchy( &self, child_id: &str, parent_id: Option<String>, ) -> Result<()>

Update the component hierarchy for bubbling

pub fn build_hierarchy_from_element( &self, element: &Element, parent_id: Option<String>, ) -> Result<()>

Build hierarchy from an element tree

pub fn take_receiver(&mut self) -> Option<UnboundedReceiver<MessageEvent>>

Get the message receiver for async processing

pub fn queue_message( &self, message: impl Message, sender_id: Option<String>, ) -> Result<()>

Add a message to the queue for batch processing

pub fn send_async( &self, message: impl Message, sender_id: Option<String>, ) -> Result<()>

Send message to async channel for processing

pub fn add_listener(&self, listener: MessageListener) -> Result<()>

Add a message listener with priority and configuration

pub fn remove_listener(&self, listener_id: &str) -> Result<bool>

Remove listeners by ID

pub async fn process_async_messages(&mut self) -> Result<()>

Process async messages from the channel

pub fn process_queue(&self) -> Result<()>

Process all queued messages

pub fn remove_element_handlers(&self, element_id: &str) -> Result<()>

Remove all handlers for a specific element (cleanup)

pub fn get_hierarchy(&self) -> Result<HashMap<String, Option<String>>>

Get hierarchy information for debugging

pub fn get_element_path(&self, element_id: &str) -> Result<Vec<String>>

Get the path from an element to the root