reactive_tui::events::messages

Trait Message

pub trait Message:
    Any
    + Send
    + Sync
    + Debug {
    // Required method
    fn clone_message(&self) -> Box<dyn Message>;

    // Provided methods
    fn type_name(&self) -> &'static str { ... }
    fn should_bubble(&self) -> bool { ... }
    fn can_prevent_default(&self) -> bool { ... }
}

Trait for messages that can be sent between components

Required Methods

fn clone_message(&self) -> Box<dyn Message>

Clone the message as a boxed trait object

Provided Methods

fn type_name(&self) -> &'static str

Get the message type name for debugging

fn should_bubble(&self) -> bool

Check if this message should bubble up the component tree

fn can_prevent_default(&self) -> bool

Check if this message can be prevented from propagating