reactive_tui::reactive

Trait ReactiveStruct

pub trait ReactiveStruct {
    // Required methods
    fn init_reactive(&mut self);
    fn reactive_state(&self) -> &ReactiveState;
    fn reactive_state_mut(&mut self) -> &mut ReactiveState;
    fn sync_to_state(&mut self) -> Result<()>;
    fn load_from_state(&mut self) -> Result<()>;
}

Trait for structs that contain reactive state and can initialize reactive watchers This provides a standard interface for reactive components

Required Methods

fn init_reactive(&mut self)

Initialize reactive watchers and setup automatic state synchronization

fn reactive_state(&self) -> &ReactiveState

Get the reactive state container for advanced state management

fn reactive_state_mut(&mut self) -> &mut ReactiveState

Get mutable access to the reactive state container

fn sync_to_state(&mut self) -> Result<()>

Sync all reactive values to the state container

fn load_from_state(&mut self) -> Result<()>

Load all reactive values from the state container