reactive_tui::widgets::animation::spring

Struct SpringConfig

pub struct SpringConfig {
    pub mass: f32,
    pub stiffness: f32,
    pub damping: f32,
    pub velocity: f32,
    pub precision: f32,
}

Configuration for spring physics calculations

Fields
mass: f32

Mass of the spring system (affects inertia)

stiffness: f32

Spring stiffness (affects oscillation frequency)

damping: f32

Damping factor (affects energy dissipation)

velocity: f32

Initial velocity of the system

precision: f32

Precision threshold for determining when spring has settled

Implementations

impl SpringConfig

pub fn new(mass: f32, stiffness: f32, damping: f32) -> Self

Create a new spring configuration

pub fn with_velocity(self, velocity: f32) -> Self

Set initial velocity

pub fn with_precision(self, precision: f32) -> Self

Set precision threshold

pub fn calculate_position(&self, time: f32, from: f32, to: f32) -> f32

Calculate spring position at given time

pub fn calculate_velocity(&self, time: f32, from: f32, to: f32) -> f32

Calculate velocity at given time

pub fn estimate_duration(&self, from: f32, to: f32) -> f32

Estimate the total duration for the spring to settle

pub fn is_settled(&self, time: f32, from: f32, to: f32) -> bool

Check if spring has settled at given time

pub fn to_easing_function(self) -> EasingFunction

Create a spring easing function that can be used with the animation system

impl SpringConfig

Preset spring configurations for common animation types

pub fn gentle() -> Self

Gentle spring with minimal overshoot

pub fn wobbly() -> Self

Wobbly spring with noticeable oscillation

pub fn stiff() -> Self

Stiff spring with quick response

pub fn slow() -> Self

Slow spring with gradual movement

pub fn bouncy() -> Self

Bouncy spring with multiple oscillations

pub fn no_overshoot() -> Self

No-overshoot spring (critically damped)