reactive_tui::widgets

Module radio

Radio button widget for single selection from a group

Provides grouped radio buttons with customizable appearance, labels, and reactive state integration.

Structs

RadioGroup

Radio button group widget

RadioGroupBuilder

Builder for creating radio groups

RadioOption

Individual radio button option

RadioState

Radio button group state

RadioStyle

Radio button styling options

Enums

RadioOrientation

Radio button group orientation

Functions

radio_group

Convenience function for creating a radio group